making a class for pyhtk.

+++ file_handling.py
@@ -5,11 +5,13 @@ import shutil
 
 
 def make_filelist(dir_target, file_txt, file_type='*'):
-    """
-    make a list of files in the directory. 
-    :param path dir_target: the directory of which a list of files will be output.
-    :param path file_txt: the text file in which a list of files in dir_target will be written.
-    :param string file_type: filetype. default is * (all). 
+    """make a list of files in the directory. 
+
+	Args:
+		dir_target (path): the directory of which a list of files will be output.
+		file_txt (path): the text file in which a list of files in dir_target will be written.
+		file_type (string): filetype. default is * (all). 
+
     """
     #listdir = os.listdir(dir_target)
     listdir = glob.glob(os.path.join(dir_target, '*.' + file_type))
@@ -20,14 +22,20 @@ def make_filelist(dir_target, file_txt, file_type='*'):
     return
 
 
-def make_new_directory(directory):
-	""" if not exist, make a new one. if exist, remove the existing directory and make a new one.
+def make_new_directory(directory, existing_dir='delete'):
+	"""make new directory. 
+
 	Args:
-		directory (path): 
+		directory (path):
+		(existing_dir) ('delete' or 'leave'): if the directory already exist, what to do. default is 'delete'.
+
 	"""
 	if os.path.exists(directory):
-		shutil.rmtree(directory)
-	os.makedirs(directory)
+		if existing_dir == 'delete':
+			shutil.rmtree(directory)
+			os.makedirs(directory)
+	else:
+		os.makedirs(directory)
 
 
 if __name__ == 'main':
